backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
langgraph>=0.0.26
langchain>=0.1.12
langchain-google-genai>=0.0.11
langchain-community>=0.0.28
python-dotenv>=1.0.1
langchain-ollama>=1.0.1
google-generativeai>=0.8.6
